+++ app/policies/building_policy.rb
@@ -2,13 +2,7 @@
 
 # Class for Building permissions.
 class BuildingPolicy < ApplicationPolicy
-  def show?
-    true
-  end
-
-  def index?
-    true
-  end
+  # We are currently exclusively inheriting behaviour from ApplicationPolicy.
 
   class Scope < Scope # rubocop:disable Style/Documentation
     def resolve