+++ app/policies/room_policy.rb
@@ -2,9 +2,7 @@
 
 # Class for Room permissions
 class RoomPolicy < ApplicationPolicy
-  def show?
-    true
-  end
+  # We are currently keeping the ability to view Rooms to admins only.
 
   class Scope < Scope # rubocop:disable Style/Documentation
     def resolve