+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 * Association of draw to draw_membership has been changed to have matching "active" states. ([#928](https://gitlab.com/yale-sdmp/vesta/issues/928)).
 * Remove code from drawStudentsUpdate to manually add students by username ([900](https://gitlab.com/yale-sdmp/vesta/issues/900)).
 * Remove stale records of DrawMemberships ([933](https://gitlab.com/yale-sdmp/vesta/issues/933)).
+* Change viewing and indexing policies for Rooms and Buildings to restrict to admin and up ([937](https://gitlab.com/yale-sdmp/vesta/issues/937)).
 
 ## v2.2.1 - 2019-02-20
 ### Added